refactor(Coin): MVC 패키지 구조를 만족시키도록 클래스 이동
